requirements.txt
six==1.13.0
websocket-client==0.56.0
